requirements.txt
# Y2AI Complete System Dependencies

# Core
requests>=2.28.0
pandas>=2.0.0
numpy>=1.24.0

# Stock Data
yfinance>=0.2.0

# Database
supabase>=2.0.0

# AI Processing
anthropic>=0.18.0

# Social Media APIs
tweepy>=4.14.0          # Twitter/X

# Scheduler
apscheduler>=3.10.0

# API Server (optional, for dashboard API)
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0

# ARGUS-1 Dependencies
feedparser>=6.0.0

# Excel reading (for Shiller data)
openpyxl>=3.1.0
xlrd>=2.0.0

# Development
python-dotenv>=1.0.0
